requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-xdist==3.5.0
pytest-html==4.1.1
python-dotenv==1.0.0
faker==20.1.0
allure-pytest==2.13.2
requests==2.31.0
Pillow==10.1.0